backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
django-ratelimit==4.1.0
django-redis==5.4.0
redis==5.0.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0

